+++ core/server/api/v0.1/settings.js
@@ -9,7 +9,7 @@ const Promise = require('bluebird'),
     models = require('../../models'),
     canThis = require('../../services/permissions').canThis,
     localUtils = require('./utils'),
-    urlService = require('../../services/url'),
+    urlService = require('../../../frontend/services/url'),
     common = require('../../lib/common'),
     settingsCache = require('../../services/settings/cache'),
     docName = 'settings';
