+++ core/server/services/slack.js
@@ -2,7 +2,7 @@ var common = require('../lib/common'),
     request = require('../lib/request'),
     imageLib = require('../lib/image'),
     urlUtils = require('../lib/url-utils'),
-    urlService = require('../services/url'),
+    urlService = require('../../frontend/services/url'),
     settingsCache = require('./settings/cache'),
     schema = require('../data/schema').checks,
     moment = require('moment'),
