+++ core/server/web/site/app.js
@@ -7,7 +7,7 @@ const config = require('../../config');
 const apps = require('../../services/apps');
 const constants = require('../../lib/constants');
 const storage = require('../../adapters/storage');
-const urlService = require('../../services/url');
+const urlService = require('../../../frontend/services/url');
 const urlUtils = require('../../lib/url-utils');
 const sitemapHandler = require('../../data/xml/sitemap/handler');
 const themeMiddleware = require('../../../frontend/services/themes').middleware;
