+++ core/test/regression/models/model_posts_spec.js
@@ -5,7 +5,7 @@ var should = require('should'),
     _ = require('lodash'),
     Promise = require('bluebird'),
     sequence = require('../../../server/lib/promise/sequence'),
-    urlService = require('../../../server/services/url'),
+    urlService = require('../../../frontend/services/url'),
     ghostBookshelf = require('../../../server/models/base'),
     models = require('../../../server/models'),
     common = require('../../../server/lib/common'),
