+++ core/test/unit/apps/amp/router_spec.js
@@ -2,7 +2,7 @@ const should = require('should'),
     sinon = require('sinon'),
     path = require('path'),
     ampController = require('../../../../frontend/apps/amp/lib/router'),
-    urlService = require('../../../../server/services/url'),
+    urlService = require('../../../../frontend/services/url'),
     helpers = require('../../../../frontend/services/routing/helpers'),
     common = require('../../../../server/lib/common'),
     testUtils = require('../../../utils'),
