+++ core/test/unit/helpers/author_spec.js
@@ -1,7 +1,7 @@
 const should = require('should'),
     sinon = require('sinon'),
     testUtils = require('../../utils'),
-    urlService = require('../../../server/services/url'),
+    urlService = require('../../../frontend/services/url'),
     helpers = require('../../../frontend/helpers');
 
 describe('{{author}} helper', function () {
