+++ core/test/unit/helpers/ghost_head_spec.js
@@ -9,7 +9,7 @@ const should = require('should'),
     models = require('../../../server/models'),
     imageLib = require('../../../server/lib/image'),
     routing = require('../../../frontend/services/routing'),
-    urlService = require('../../../server/services/url'),
+    urlService = require('../../../frontend/services/url'),
     helpers = require('../../../frontend/helpers'),
     proxy = require('../../../frontend/helpers/proxy'),
     settingsCache = proxy.settingsCache;
