+++ core/test/unit/helpers/tags_spec.js
@@ -1,7 +1,7 @@
 const should = require('should'),
     sinon = require('sinon'),
     testUtils = require('../../utils'),
-    urlService = require('../../../server/services/url'),
+    urlService = require('../../../frontend/services/url'),
     models = require('../../../server/models'),
     helpers = require('../../../frontend/helpers');
