+++ core/test/unit/models/post_spec.js
@@ -5,7 +5,7 @@ const sinon = require('sinon');
 const Promise = require('bluebird');
 const testUtils = require('../../utils');
 const knex = require('../../../server/data/db').knex;
-const urlService = require('../../../server/services/url');
+const urlService = require('../../../frontend/services/url');
 const schema = require('../../../server/data/schema');
 const models = require('../../../server/models');
 const common = require('../../../server/lib/common');
