+++ core/test/unit/services/routing/controllers/collection_spec.js
@@ -4,7 +4,7 @@ const should = require('should'),
     common = require('../../../../../server/lib/common'),
     security = require('../../../../../server/lib/security'),
     themeService = require('../../../../../frontend/services/themes'),
-    urlService = require('../../../../../server/services/url'),
+    urlService = require('../../../../../frontend/services/url'),
     controllers = require('../../../../../frontend/services/routing/controllers'),
     helpers = require('../../../../../frontend/services/routing/helpers');
