+++ core/test/unit/services/rss/generate-feed_spec.js
@@ -3,7 +3,7 @@ var should = require('should'),
     _ = require('lodash'),
     testUtils = require('../../../utils'),
     urlUtils = require('../../../utils/urlUtils'),
-    urlService = require('../../../../server/services/url'),
+    urlService = require('../../../../frontend/services/url'),
     generateFeed = require('../../../../server/services/rss/generate-feed');
 
 describe('RSS: Generate Feed', function () {
