+++ core/test/unit/services/url/UrlGenerator_spec.js
@@ -4,7 +4,7 @@ const should = require('should');
 const nql = require('@nexes/nql');
 const sinon = require('sinon');
 const urlUtils = require('../../../../server/lib/url-utils');
-const UrlGenerator = require('../../../../server/services/url/UrlGenerator');
+const UrlGenerator = require('../../../../frontend/services/url/UrlGenerator');
 
 describe('Unit: services/url/UrlGenerator', function () {
     let queue, router, urls, resources, resource, resource2;
