+++ core/test/utils/index.js
@@ -19,7 +19,7 @@ var Promise = require('bluebird'),
     schemaTables = Object.keys(schema),
     models = require('../../server/models'),
     urlUtils = require('../../server/lib/url-utils'),
-    urlService = require('../../server/services/url'),
+    urlService = require('../../frontend/services/url'),
     routingService = require('../../frontend/services/routing'),
     settingsService = require('../../server/services/settings'),
     settingsCache = require('../../server/services/settings/cache'),
